Availability of 0.1.0 alpha-5

+++ scirs2-core/examples/memory_mapped_chunks.rs
@@ -144,7 +144,7 @@ fn aggregate_example(temp_dir: &Path) -> Result<(), Box<dyn std::error::Error>> 
         .unwrap();
     let global_sum = chunk_stats
         .iter()
-        .map(|&(_, _, sum, _, _)| sum as i64)
+        .map(|&(_, _, sum, _, _)| sum)
         .sum::<i64>();
     let global_count = chunk_stats
         .iter()
